requirements.txt
django
google-cloud-bigquery
python-dotenv
PyJWT
cryptography
google-generativeai